Refactor out esbuild configs; add zip plugin and typechecking
