+++ package.json
@@ -46,9 +46,11 @@
     "webextension-polyfill": "^0.12.0"
   },
   "devDependencies": {
+    "@jgoz/esbuild-plugin-typecheck": "^4.0.1",
     "@tailwindcss/forms": "^0.5.7",
     "@testing-library/jest-dom": "^6.4.6",
     "@testing-library/react": "^16.0.0",
+    "@types/archiver": "^6.0.2",
     "@types/jest": "^29.5.12",
     "@types/node": "^20.14.10",
     "@types/react": "^18.3.3",
@@ -58,6 +60,7 @@
     "@types/zip-webpack-plugin": "^3.0.6",
     "@typescript-eslint/eslint-plugin": "^7.16.0",
     "@typescript-eslint/parser": "^7.16.0",
+    "archiver": "^7.0.1",
     "autoprefixer": "^10.4.19",
     "clean-webpack-plugin": "^4.0.0",
     "copy-webpack-plugin": "^12.0.2",
