+++ src/popup/index.html
@@ -5,7 +5,7 @@
     <title>Web Monetization Extension</title>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="./popup.css" />
-    <script src="./popup.js" defer></script>
+    <script defer src="./popup.js"></script>
   </head>
   <body>
     <div id="popup-container"></div>
