+++ tsconfig.json
@@ -24,5 +24,5 @@
     }
   },
   "include": ["./src/**/*", "./jest.config.ts", "./jest.setup.ts"],
-  "exclude": ["dist", "dev", "temp"]
+  "exclude": ["dist", "dev"]
 }
